Made cron.php only accessible by admins, cronjob.php is what should be grabbed by the cronjob.
